requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
aiosqlite=0.19.0=pypi_0
anyio=3.6.2=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.2.0=pypi_0
autocommand=2.2.2=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.2=pypi_0
bidict=0.22.1=pypi_0
black=23.1.0=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
blis=0.7.10=pypi_0
ca-certificates=2022.10.11=h06a4308_0
cachelib=0.10.2=pypi_0
catalogue=2.0.9=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=3.0.1=pypi_0
cheroot=10.0.0=pypi_0
cherrypy=18.8.0=pypi_0
click=8.1.3=pypi_0
confection=0.1.1=pypi_0
cryptography=41.0.3=pypi_0
cycler=0.11.0=pypi_0
cymem=2.0.7=pypi_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.7=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
docx=0.2.4=pypi_0
en-core-web-sm=3.6.0=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.1.0=pypi_0
fastjsonschema=2.16.2=pypi_0
feedparser=6.0.10=pypi_0
filelock=3.9.0=pypi_0
flask=2.2.5=pypi_0
flask-session=0.5.0=pypi_0
flask-socketio=5.3.6=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
frozenlist=1.3.3=pypi_0
future=0.18.2=py37_1
gevent=22.10.2=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
greenlet=3.0.3=pypi_0
gunicorn=21.2.0=pypi_0
h11=0.14.0=pypi_0
huggingface-hub=0.12.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=5.2.0=pypi_0
importlib-resources=5.10.2=pypi_0
inflect=6.0.5=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.2=pypi_0
ipympl=0.9.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.4=pypi_0
itsdangerous=2.1.2=pypi_0
jaraco-collections=4.2.0=pypi_0
jaraco-context=4.3.0=pypi_0
jaraco-functools=3.7.0=pypi_0
jaraco-text=3.11.1=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-ai=0.2.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.23.5=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.5=pypi_0
kiwisolver=1.4.4=pypi_0
langcodes=3.3.0=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
line-profiler=4.0.3=pypi_0
lxml=4.9.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
more-itertools=9.1.0=pypi_0
multidict=6.0.4=pypi_0
murmurhash=1.0.9=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclassic=0.5.1=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.9=pypi_0
nbformat=5.7.3=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nltk=3.8.1=pypi_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openai=0.27.0=pypi_0
openssl=1.1.1s=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.0=pypi_0
pathtools=0.1.2=pypi_0
pathy=0.10.2=pypi_0
patsy=0.5.6=pypi_0
pdfminer-six=20221105=pypi_0
pdfminer3k=1.3.4=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py37h06a4308_0
pipreqs=0.4.13=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.0.0=pypi_0
plotly=5.12.0=pypi_0
pluggy=1.0.0=pypi_0
ply=3.11=pypi_0
portend=3.1.0=pypi_0
preshed=3.0.8=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.36=pypi_0
protobuf=4.21.12=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.12=pypi_0
pygments=2.14.0=pypi_0
pyinflect=0.5.1=pypi_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
pytest=7.2.1=pypi_0
python=3.7.15=h7a1cb2a_1
python-dateutil=2.8.2=pypi_0
python-engineio=4.8.0=pypi_0
python-socketio=5.10.0=pypi_0
pytorch=1.12.1=cpu_py37hb1f1ab4_1
pytz=2022.7.1=pypi_0
pyyaml=6.0=py37h5eee18b_1
pyzmq=25.0.0=pypi_0
qtconsole=5.4.0=pypi_0
qtpy=2.3.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
send2trash=1.8.0=pypi_0
sentry-sdk=1.13.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
sgmllib3k=1.0.0=pypi_0
simple-websocket=1.0.0=pypi_0
simplejson=3.19.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.3.0=pypi_0
smmap=5.0.0=pypi_0
snakeviz=2.1.1=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.4=pypi_0
sqlite=3.40.1=h5082296_0
srsly=2.4.7=pypi_0
statsmodels=0.13.5=pypi_0
tempora=5.3.0=pypi_0
tenacity=8.1.0=pypi_0
terminado=0.17.1=pypi_0
thinc=8.1.12=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
tomli=2.0.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.26.1=pypi_0
typed-ast=1.5.4=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.4.0=py37h06a4308_0
typing_extensions=4.4.0=py37h06a4308_0
urllib3=1.26.14=pypi_0
wandb=0.13.9=pypi_0
wasabi=1.1.2=pypi_0
wcwidth=0.2.6=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.0=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.5=pypi_0
wsproto=1.2.0=pypi_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarg=0.1.9=pypi_0
yarl=1.8.2=pypi_0
zc-lockfile=3.0.post1=pypi_0
zipp=3.11.0=pypi_0
zlib=1.2.13=h5eee18b_0
zope-event=5.0=pypi_0
zope-interface=6.1=pypi_0
